requirements_Version2.txt
cdsapi
xarray
pandas
numpy
dask[complete]
xesmf
netCDF4
pytorch
torchvision
scikit-learn
streamlit
fastapi
uvicorn
aiohttp
pyyaml
matplotlib
seaborn